Fixed Year dropdown bug
As noted by the professor:
"There is a change required to the YearDropDownList from previous labs
to make the custom validator for the credit card expiry date work
correctly. Prior to running the code to populate the YearDropDownList,
the drop down list should contain one item with Text property "Year"
and Value property "0". If it does not, then you will get an exception
when the validator tries to convert the value "Year" to an integer when
validating the credit card expiry date. The code that you were given in
the security lab to populate the drop down lists assumes that the
YearDropDownList already has this one element, so there is nothing to
change in the code."